requirements.txt
google-cloud-vision
piexif
